fix: Loading tilesets with spacing (#368)
* Added a test map for tilesets with spacing in them
* Fixed loading of maps with spacing

Explanation on the fix: 
The number of columns and rows were not being calculated correctly, which threw off which sprite it was pulling from the sprite sheet.
If we had a tileset with 3 tiles, 16px wide and 1px spacing, the width of the image would be 50px. We don't get clean numbers if we do `50 / (16 + 1)` because there are only spaces between the tiles, but not before/after the first/last tiles. So we can just add the spacing to the height/width, and then divide by the tile width (or height) + spacing.
Ex:
```
(50px + 1px spacing) / (16px width + 1px spacing)
51 / 17 = 
3 tiles
```

**Before**
![image](https://user-images.githubusercontent.com/4239808/170203047-c0a1b839-d15a-44c5-9c81-caddd755cd1b.png)

**After**
![image](https://user-images.githubusercontent.com/4239808/170203266-4dd9ede3-dcb3-4d7d-b78f-da6a9c1c888a.png)

**Tiled Map**
![image](https://user-images.githubusercontent.com/4239808/170203447-370719da-cc49-4fd8-b083-87244a9f317b.png)